APPLIED (dm): test/py: capture the entire U-Boot version at boot
The existing regex simply ensures that the captured version string doesn't
go past the end of a line. We really want to grab as much as possible. Do
this by explicitly including a ) character at the end of the regex to
match the last character of the version test.

Signed-off-by: Stephen Warren <swarren@nvidia.com>
---
v2: Also remove some now-redundant code. The signon regex now never
matches the text ", Build: ...", so there's no need for explicit code to
chop it out.